--- lib/core/scene.dart
@@ -1,13 +1,0 @@
-import 'package:npdart/core/stage.dart';
-import 'package:npdart/core/state.dart';
-
-class Scene {
-  final Future<NovelStateSnapshot> Function(
-      Stage stage, NovelStateSnapshot state) script;
-  final String? description;
-
-  Scene({
-    required this.script,
-    this.description,
-  });
-}
