+++ lib/npdart.dart
@@ -1,1 +1,4 @@
-export 'core/core.dart';
+///A minimal game engine for creating visual novels.
+library;
+
+export 'src/core/core.dart';
