+++ lib/src/core/choice.dart
