+++ lib/src/core/core.dart
@@ -6,3 +6,5 @@ export 'save.dart';
 export 'scene.dart' show Scene;
 export 'tree.dart' show Tree;
 export 'verse.dart' show Verse;
+export 'stage.dart' show Stage;
+export 'event.dart';
