+++ lib/src/core/event.dart
