+++ lib/src/core/mouse.dart
