+++ lib/src/core/verse.dart
@@ -1,7 +1,9 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 
-///Информация, выводимая в текстовом блоке сцены
+///A phrase that can be printed to the textbox of a scene.
+///
+///Note that a [color] will be applied to a [header], but not to a [string].
 class Verse extends Equatable {
   final String header;
   final String string;
