+++ lib/src/widgets/painting/spritepainter.dart
