+++ lib/src/widgets/ui/autoskip.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:npdart/core/core.dart';
-import 'package:npdart/core/event.dart';
-import 'package:npdart/core/stage.dart';
-import 'package:npdart/core/state.dart';
+import 'package:npdart/src/core/core.dart';
+import 'package:npdart/src/core/event.dart';
+import 'package:npdart/src/core/stage.dart';
+import 'package:npdart/src/core/state.dart';
 
 class AutoskipDialog extends StatefulWidget {
   const AutoskipDialog({
